Make user flow overflow bucket collision-safe and expand coverage
Track the $others overflow bucket as a distinct internal identity
(nodeRef{id, others}) in GroupUserFlowResult and expose it via a new
UserFlowNode.is_others field, so a real node whose id is literally
"$others" never merges with the synthetic bucket. The bucket's emitted
id is disambiguated when a real surviving node already uses that id, so
link endpoints stay unambiguous; clients identify the bucket by the flag.

- Drop UserFlowNode.label (was always equal to id).
- Drop the unused GroupBy parameter from the session grouping helpers,
  renaming them userFlowSessionGroupKeyColumn / userFlowNonEmptySessionKeyCond;
  only GROUP_BY_SESSION exists, so a future GROUP_BY_USER must add its column.
- Scan the count as uint64 in the dashboards render-test mock to match
  QueryUserFlow, and document the int64 narrowing in the executor.

Coverage added: overflow summation, the user-flow CEL validation rules,
the max_hops default, real-"$others" non-merge, ClickHouse-backed
$others/truncation and time-ordering, the scope baseline, and handler
link values. docs/architecture/user-flow.md reconciled with the shipped code.

+++ internal/core/dashboards/query_test.go
@@ -97,7 +97,9 @@ func (r *userFlowRows) Scan(dest ...any) error {
 	r.idx++
 	*dest[0].(*string) = row[0].(string)
 	*dest[1].(*string) = row[1].(string)
-	*dest[2].(*int64) = row[2].(int64)
+	// QueryUserFlow scans the count into a uint64 (ClickHouse count(DISTINCT) is
+	// UInt64) before narrowing to int64; the mock must match that destination type.
+	*dest[2].(*uint64) = uint64(row[2].(int64))
 	return nil
 }
